// Features/reports/Types/IVacationForReport.ts
export interface IVacationForReport {
    identifier: string;
    idEmployee: number;
    employeeName: string;
    startDate: string;
    endDate: string;
    isPaid: string;
    paid: string;
    payDate: string;
    enjoymentDay: number;
    absenteeism: number;
    dayToPay: number;
    payrollName: string;
    payrollNumber: number;
    idPosition: number;
    idPayrollPay: number;
    idDepartment: number;
    position: string;
    department: string;
    amount: number;
    idConcept: number;
    conceptCode: string;
    concept: string;
    idCompany: number;
}